Complete User Registration Flow
Tweaked the User Registration flow so users can register, login and are
automatically redirected to the dashboard once authenticated
