+++ client/register.html
@@ -1,0 +1,34 @@
+<html>
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <title>Ariadne</title>
+  <link href="https://maxcdn.bootstrapcdn.com/bootswatch/3.3.6/cosmo/bootstrap.min.css" rel="stylesheet" integrity="sha384-OiWEn8WwtH+084y4yW2YhhH6z/qTSecHZuk/eiWtnvLtU+Z8lpDsmhOKkex6YARr" crossorigin="anonymous">
+</head>
+<body>
+  <main id="root">
+    <div class="container">
+      <div class="page-header">
+        <h1><a href="/">Ariadne</a> <small>Create an Account</small></h1>
+      </div>
+      <form action="" method="post">
+        <div class="form-group">
+          <label for="username">Username</label>
+          <input type="text" class="form-control" name="username" placeholder="Username">
+        </div>
+        <div class="form-group">
+          <label for="displayName">Display Name</label>
+          <input type="text" class="form-control" name="displayName" placeholder="Display Name">
+        </div>
+        <div class="form-group">
+          <label for="password">Password</label>
+          <input type="password" class="form-control" name="password" placeholder="Password">
+        </div>
+        <input type="submit" class="btn btn-success" value="Create Account">
+      </form>
+      <a href="/register">Don't have an account? Sign up here.</a>
+    </div>
+  </main>
+
+</body>
+</html>
